requirements.txt
django==6.0
python-dotenv==1.2.1
pillow==12.0.0
google-genai==1.56.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2==2.9.11
psycopg2-binary==2.9.11